backend/requirements.txt
# Gamuda PM Platform — Python dependencies
# Install with: pip install -r requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9      # Required for OAuth2PasswordRequestForm
